refactored edit profile to components
